+++ listings/ch17-async-await/listing-17-26/src/main.rs
@@ -4,11 +4,11 @@ fn main() {
     trpl::block_on(async {
         // ANCHOR: here
         let slow = async {
-            trpl::sleep(Duration::from_secs(5)).await;
+            trpl::sleep(Duration::from_millis(100)).await;
             "I finished!"
         };
 
-        match trpl::timeout(Duration::from_secs(2), slow).await {
+        match timeout(slow, Duration::from_millis(10)).await {
             Ok(message) => println!("Succeeded with '{message}'"),
             Err(duration) => {
                 println!("Failed after {} seconds", duration.as_secs())
